+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Don't prune followed conversations.
 - Include current user's own reactions in post reaction list.
+- Do not attempt to sync conversation when root is deleted.
 
 ## [4.16.1] - 2026-01-07